requirements.txt
selenium
langchain
langchain-openai
langchain-anthropic
langgraph
pydantic
python-dotenv
bs4
